Fix `kaggle.json` moving command
`kaggle.json` is the file that stores Kaggle access token.
To download models via 'kagglehub', the file must be stored in
the directory of `~/.kaggle/`. But the error was caused since
the directory to make was given explicitly. `~` implies home directory,
but if the directory string is given expicitly by wrapping with
quotation marks(in this case, like `"~/.kaggle"`), it will find the
directory that is named as `~` instead of the home directory - which was
intended to be.